Add some basic cargo workspaces for the project structure
Workspaces allow to aggregate several crates in a project structure to
build them in a shared environment that persists all build artifacts in
the `target/` directory in the project's root directory.

This ensures that no duplications occur and a standardized structure can
be used for the entire project.

See also: https://doc.rust-lang.org/book/ch14-03-cargo-workspaces.html