Avoid unneeded rebuilds for frontend crates
In our linker-flags for the frontend build we have different stack-sizes
given which results in different binaries. Therefore the compiled crates
are slightly different (comparing those from logic to graphics). That
caused almost full rebuilds whenever one of the crates should be recompiled.

To avoid any clashes (during build or runtime), the target/ directories
need to be separated here (we build with different linker flags and those
are the only crates for wasm32-unknown-unknown so we're not loosing
cargo-workspace benefits here).

Closes #54